move to hunter plugin, color is now empty trap hunter color
